refactor: add all types to one file
