--- src/types/LangConstants.ts
@@ -1,11 +1,0 @@
-interface LangConstants {
-  langButton: string;
-  altLogo: string;
-  textCreated: string;
-  welcomePageLink: string;
-  signInBtn: string;
-  signUpBtn: string;
-  toMainBtn: string;
-}
-
-export default LangConstants;
