--- src/types/LangContext.ts
@@ -1,11 +1,0 @@
-import LangConstants from '@/types/LangConstants';
-
-interface LangContext {
-  en: LangConstants;
-  ru: LangConstants;
-  pageLang: string;
-  setPageLang: (lang: string) => void;
-  getConstants: () => LangConstants;
-}
-
-export default LangContext;
