--- src/types/PropsAuthButton.ts
@@ -1,5 +1,0 @@
-interface PropsAuthButton {
-  type: 'sign-in' | 'sign-up' | 'to-main';
-}
-
-export default PropsAuthButton;
