+++ src/types/PropsButtonComponent.ts
@@ -1,7 +1,1 @@
-interface StylesOfButton {
-  button: string;
-  textButtonStyle: string;
-  textButton: string;
-}
 
-export default StylesOfButton;
